Fixed two crashes
Crash 1 (at network_downloader.cpp:61): This crash happened when the app would attempt to copy the vector data to put it into the map. This fix is technically more of a workaround, as it prevents the vector from getting copied, and instead converts it into a rvalue by std::moving it, which means it doesn't need to be copied and can directly be put into the map

Crash 2 (at draw.cpp:123): This crash would happen when the code attempted to set c2d_image->subtex->top to 1.0. This happened because Draw_c2d_image_init didn't allocate the right amount of memory. Instead of allocating the sizeof(Tex3DS_SubTexture) and sizeof(C3D_Tex), it allocated the size of their pointers, meaning it just allocated 4 bytes two times and not the size of the structure. This meant that in linear space, there would be 8 allocated bytes in a row, which fit the width and height, which are both u16 values, filling 4 bytes, and the left variable, which is a 32-bit float, which is 4 bytes, and fills the allocated RAM. When attempting to set top, which happens at an offset of exactly 8 bytes, it wanders off into unallocated memory, which is known as undefined behavior and leads to random segfaults.